requirements_test.txt
pytest>=8.0
pytest-asyncio>=0.23
pytest-homeassistant-custom-component>=0.13
pytest-cov>=5.0
aiohttp>=3.9
voluptuous>=0.14
